test-automation/IELTS/requirements.txt
mysql_connector_repackaged==0.3.1
pytest==8.3.3
Requests==2.32.3
selenium==4.25.0
sql==2022.4.0
webdriver_manager==4.0.2
